[Hackability Refactor] Move export and generate under torchchat/ (#1089)
