+++ torchchat/usages/openai_api.py
@@ -16,8 +16,8 @@ from typing import Any, Dict, List, Optional, Union
 import torch
 
 from torchchat.utils.build_utils import device_sync
-from download import is_model_downloaded, load_model_configs
-from generate import Generator, GeneratorArgs
+from torchchat.cli.download import is_model_downloaded, load_model_configs
+from torchchat.generate import Generator, GeneratorArgs
 
 
 """Dataclasses defined around the objects used the OpenAI API Chat specification.
